refactor(topology): unify to single topology.json; rename ToOut → ToReadGate per specific-name rule
- main.go default --topology flag changed from topologies/line.json to topology.json
- topology.json: rename all ToOut port names/sourceHandles to ToReadGate (matches InputNode struct field)
- nodes/Wiring/loader.go: specEdge.Lane int → float64 to accept editor-generated float lane values
- topologies/line.json deleted; topologies/ dir removed
- handoff.md: remove stale line.json open-bug section, update go run . reference

+++ nodes/Wiring/loader.go
@@ -43,7 +43,7 @@ type NodeData struct {
 type specEdge struct {
 	Label        string `json:"label"        wire:"prop,optional,tsType:string"`
 	ValueLabel   string `json:"valueLabel"   wire:"prop,optional,tsType:string"`
-	Lane         int    `json:"lane"         wire:"prop,optional,tsType:number"`
+	Lane         float64 `json:"lane"         wire:"prop,optional,tsType:number"`
 	ArrowStyle   string `json:"arrowStyle"   wire:"prop,optional,tsType:ArrowStyle"`
 	Concurrent   *bool  `json:"concurrent"   wire:"prop,optional,tsType:boolean"`
 	Kind         string `json:"kind"         wire:"prop,required,tsType:EdgeKind"`
